Added NOROOT.md
I found out how to import Steam accounts from an un-rooted Android phone, although the process is much longer and complicated, it's still possible. It need more testing though, as I have only tested it with my phone (and it does work). If you decide to use it, I will try to create a script file to mostly automate it.